requirements.txt
langchain
langchain-community
transformers
sentence-transformers
chromadb
unstructured
pdfminer.six
tqdm
python-dotenv
